refactor: move plugin unit test
