+++ plugins/hammer/src/__test__/hammer-mock.js
