+++ plugins/q/package.json
@@ -16,7 +16,7 @@
   "scripts": {
     "build": "cross-env NODE_ENV=production rollup --config ../../rollup.config.js",
     "build:watch": "yarn run build -w",
-    "lint": "eslint src test",
+    "lint": "eslint src",
     "prepublishOnly": "rm -rf dist && npm run build"
   },
   "devDependencies": {
