+++ plugins/q/src/data/__test__/field.spec.js
@@ -1,6 +1,6 @@
 // import * as picasso from '../../../../src/index';
 
-import qField from '../../src/data/field';
+import qField from '../field';
 
 describe('q-field', () => {
   function struct(mode, fe, type) {
