+++ plugins/q/src/formatter/__test__/formatter.spec.js
@@ -1,4 +1,4 @@
-import { createFromMetaInfo } from '../../../src/formatter';
+import { createFromMetaInfo } from '..';
 
 const qLocaleInfo = {
   qDecimalSep: ',',
