+++ plugins/q/src/scales/qTime/__test__/qTime.spec.js
@@ -1,4 +1,4 @@
-import qTime from '../../../../src/scales/qTime';
+import qTime from '..';
 import tickGen from './tickGen';
 
 describe('Scale - qTime', () => {
